+++ opt-env-conf-example/app/Main.hs
@@ -1,0 +1,6 @@
+module Main (main) where
+
+import OptEnvConf.Example (exampleMain)
+
+main :: IO ()
+main = exampleMain