+++ opt-env-conf-example/test/OptEnvConf/ExampleSpec.hs
@@ -1,0 +1,13 @@
+{-# LANGUAGE TypeApplications #-}
+
+module OptEnvConf.ExampleSpec where
+
+import OptEnvConf.Example
+import OptEnvConf.Test
+import Test.Syd
+
+spec :: Spec
+spec = do
+  settingsLintSpec @Instructions
+  goldenSettingsReferenceDocumentationSpec @Instructions "test_resources/documentation.txt" "opt-env-conf-example"
+  goldenSettingsNixOptionsSpec @Instructions "options.nix"